_#2 Make input from console not cut off when receiving a message
The problem to achieve this was to cache all the bytes written
at the time a message arrives. A way to solve this is getting
one char at the time and save that in other place (in this
case using a channel).

Another channel is used to know when a new message has been
received, so can clean the line, print the message and print
the text the user has been writing at the moment.

Also:
  - Add constants
  - Close resources

Signed-off-by: Denis Reinoso <reinoso.denis@gmail.com>